+++ ui/mods/maps/maplist.js
@@ -61,11 +61,11 @@ cShareSystems.load_pas("eXodus 1v1", [
 	"coui://ui/mods/maps/systems/tectonic.pas",
 	"coui://ui/mods/maps/systems/theclassybarren.pas",
 	"coui://ui/mods/maps/systems/thermopylae.pas",
-	"coui://ui/mods/maps/systems/thewall.pas",
-	"coui://ui/mods/maps/systems/thief's_pass.pas",
 	"coui://ui/mods/maps/systems/the_faceoff.pas",
 	"coui://ui/mods/maps/systems/the_forgery.pas",
 	"coui://ui/mods/maps/systems/the_reef.pas",
+	"coui://ui/mods/maps/systems/thewall.pas",
+	"coui://ui/mods/maps/systems/thief's_pass.pas",
 	"coui://ui/mods/maps/systems/toisengard.pas",
 	"coui://ui/mods/maps/systems/tundra.pas",
 	"coui://ui/mods/maps/systems/venetzia.pas",
@@ -78,18 +78,17 @@ cShareSystems.load_pas("eXodus 2v2", [
 	"coui://ui/mods/maps/systems/fields_of_isis.pas",
 	"coui://ui/mods/maps/systems/frozenoreoyoghurt.pas",
 	"coui://ui/mods/maps/systems/gsystem_vs2v2.pas",
-	"coui://ui/mods/maps/systems/icy_v17.pas",
 	"coui://ui/mods/maps/systems/incontrovertible.pas",
 	"coui://ui/mods/maps/systems/kinetics_multi_4v.pas",
 	"coui://ui/mods/maps/systems/lockery.pas",
 	"coui://ui/mods/maps/systems/seton's_grip.pas",
+	"coui://ui/mods/maps/systems/icy_v17.pas",
 	"coui://ui/mods/maps/systems/the_t3mple_v2.pas",
 	"coui://ui/mods/maps/systems/elrose_v18.pas",
 	"coui://ui/mods/maps/systems/t-_kure atoll.pas",
 	"coui://ui/mods/maps/systems/tandem.pas",
 	"coui://ui/mods/maps/systems/the_forgery.pas",
 	"coui://ui/mods/maps/systems/the_t3mple_v2.pas",
-	"coui://ui/mods/maps/systems/trials.pas",
 	"coui://ui/mods/maps/systems/tropicanhoneydonut.pas",
 	"coui://ui/mods/maps/systems/zetox.pas"
 ]);
@@ -104,8 +103,7 @@ cShareSystems.load_pas("eXodus 3v3+", [
 	"coui://ui/mods/maps/systems/seton's_grip.pas",
 	"coui://ui/mods/maps/systems/t-_rain_system.pas",
 	"coui://ui/mods/maps/systems/twbaby.pas",
-	"coui://ui/mods/maps/systems/war_of the worlds.pas",
-	"coui://ui/mods/maps/systems/worlds_collide.pas"
+	"coui://ui/mods/maps/systems/war_of the worlds.pas"
 ]);
 cShareSystems.load_pas("eXodus FFA", [
 	"coui://ui/mods/maps/systems/newzealandsbaby.pas",