+++ ui/mods/maps/systems/wadiya_no_crater.pas
@@ -1,0 +1,92 @@
+{
+    "name": "Wadiya - No Crater",
+    "description":"Wadiya without the craters.  (Fixed - really without craters)",
+    "creator":"AndreasG",
+    "version":"1.2",
+    "planets": [
+        {
+            "name": "Nostalgica - No Crater",
+            "mass": 5000,
+            "position_x": 37500,
+            "position_y": -6200,
+            "velocity_x": 18.708784103393555,
+            "velocity_y": 113.15800476074219,
+            "required_thrust_to_move": 3,
+            "starting_planet": true,
+            "respawn": false,
+            "start_destroyed": false,
+            "min_spawn_delay": 0,
+            "max_spawn_delay": 0,
+            "planet": {
+                "seed": 356356192,
+                "radius": 550,
+                "heightRange": 0,
+                "waterHeight": 0,
+                "waterDepth": 100,
+                "temperature": 0,
+                "metalDensity": 100,
+                "metalClusters": 92,
+                "metalSpotLimit": -1,
+                "biomeScale": 50,
+                "biome": "moon",
+                "symmetryType": "none",
+                "symmetricalMetal": false,
+                "symmetricalStarts": false,
+                "numArmies": 2,
+                "landingZonesPerArmy": 0,
+                "landingZoneSize": 0
+            },
+            "planetCSG": [
+                {
+                    "no_features": false,
+                    "mergeable": true,
+                    "pathable": true,
+                    "flooded": false,
+                    "twinId": 0,
+                    "scale": [
+                        0.020000386983156204,
+                        0.020000386983156204,
+                        0.020000386983156204
+                    ],
+                    "weightScale": [
+                        1,
+                        1,
+                        1
+                    ],
+                    "weight": [
+                        0,
+                        0,
+                        0,
+                        0
+                    ],
+                    "weightHard": false,
+                    "height": 550.8327026367188,
+                    "position": [
+                        -150.8565673828125,
+                        215.0583953857422,
+                        484.15789794921875
+                    ],
+                    "mirrored": false,
+                    "proj": "BP_Bend",
+                    "transform": [
+                        0.019202005118131638,
+                        0.0011381588410586119,
+                        -0.00547750573605299,
+                        -150.85655212402344,
+                        0.0011381588410586119,
+                        0.018377847969532013,
+                        0.007808633614331484,
+                        215.05838012695312,
+                        0.00547750573605299,
+                        -0.007808633614331484,
+                        0.017579466104507446,
+                        484.15789794921875
+                    ],
+                    "op": "BO_Subtract",
+                    "rotation": 0,
+                    "spec": "/pa/terrain/moon/brushes/moon_crater_02.json"
+                }
+            ]
+        }
+    ]
+}