+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Fix for Elijahs Path SHARED missing a metal spot near one of the spawns thanks to Grand Homie
   - Clarified that Mots Recipe For Disaster in 2v2 and 3v3 are different
   - Removed FFA from map names due to new tab system
+  - Fix map sorting
   - New 1v1 maps
     - Dust 1.0
     - Solitude 1.0